fix room latest evt hook
